Adding Info - In Progress
Changing how info is kept and displayed, in order to figure out what is
wrong (currerntly in progress, not complete). Creating global variables
to see how far off the robot is from the desired area, needed to be
checking if logic is correct (since the robot does not appear to desire
to stop).